requirements.txt
pytest~=7.2.0
selenium~=4.7.2
allure-pytest~=2.12.0
Faker~=15.3.4
randomuser~=1.6
SQLAlchemy~=1.4.45